fix: resolve active compilation blockers and create status guide
Resolves several active compilation and borrow-checker failures:
- Removes duplicate constructor and implements Hash trait on Version in sigpkg/mod.rs.
- Fixes return syntax mismatch in ContextWindowPruner::new inside ai/orchestrator.rs.
- Corrects unclosed klib/paging.rs helper block, wrapping unit tests cleanly.
- Deletes duplicate/overlapping merge blocks in vulnerability.rs and capability.rs.
- Moves non-trait methods to concrete SimpleStorageDriver block in driver/framework.rs.
- Casts AtomicUsize loaded values to u32 before transmuting across oci_runtime.rs,
  runtime.rs, device/manager.rs, and driver/device.rs to resolve transmute size mismatch.
- Adds missing field initializations in SimpleCompositor and ShellRepl constructors, and
  defines a mock AgentAutomationEngine.
- Updates ALGORITHMS_STATUS.md to clean up pre-existing merge conflicts.
- Creates definitive WHAT_IS_WORKING_AND_NOT_WORKING.md status guide.

+++ src/container/oci_runtime.rs
@@ -1,6 +1,3 @@
-#![no_std]
-#![no_main]
-
 use core::mem;
 /// OOP-based Container Runtime Support for SigmaOS
 /// Based on Ideas-999-Structured: Core System Item 17
@@ -10,7 +7,7 @@ use core::sync::atomic::{AtomicUsize, Ordering};
 pub type ContainerID = usize;
 
 #[repr(C)]
-#[derive(Debug, Clone, Copy)]
+#[derive(Debug, Clone, Copy, PartialEq, Eq)]
 pub enum ContainerState {
     Created = 0,
     Running = 1,
@@ -72,7 +69,7 @@ impl Container for SimpleContainer {
         &self.name[..len]
     }
     fn state(&self) -> ContainerState {
-        unsafe { core::mem::transmute(self.state.load(Ordering::SeqCst)) }
+        unsafe { core::mem::transmute(self.state.load(Ordering::SeqCst) as u32) }
     }
 
     fn start(&mut self) -> Result<(), ContainerError> {
