+++ src/device/manager.rs
@@ -67,7 +67,7 @@ impl Device for SimpleDevice {
         &self.name[..len]
     }
     fn device_class(&self) -> DeviceClass {
-        unsafe { core::mem::transmute(self.device_class.load(Ordering::SeqCst)) }
+        unsafe { core::mem::transmute(self.device_class.load(Ordering::SeqCst) as u32) }
     }
 
     fn initialize(&mut self) -> Result<(), DeviceError> {
